ØSLKDJflkdsj

+++ VG2/Matematikk R2/VektormatteA.py
@@ -1,9 +1,9 @@
 import numpy as np
 
 print("------ Vektor 1 -------")
-u = np.array([float(input("X: ")), float(input("Y: "))])
+u = np.array([int(input("X: ")), int(input("Y: "))])
 print("------ Vektor 2 -------")
-v = np.array([float(input("X: ")), float(input("Y: "))])
+v = np.array([int(input("X: ")), int(input("Y: "))])
 
 utregning1 = u + v
 utregning2 = u - v
